# openapi.yaml
openapi: 3.0.3
info:
  title: Inactu Control API
  version: 0.1.0
  description: Temporary control-plane API scaffold for Inactu.
servers:
  - url: http://localhost:8080
paths:
...
  /v1/packages:
    get:
      summary: List packages for the authenticated owner
      operationId: getPackages
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Package list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPackagesResponse"
        "401":
          description: Missing or invalid user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create a package under the authenticated owner
      operationId: postPackages
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackageRequest"
      responses:
        "200":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePackageResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Package already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/packages/{package}/versions:
    get:
      summary: List versions for a package
      operationId: getPackageVersions
      parameters:
        - in: path
          name: package
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Version list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPackageVersionsResponse"
        "401":
          description: Missing or invalid user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Publish a package version
      operationId: postPackageVersion
      parameters:
        - in: path
          name: package
          required: true
          schema:
            type: string
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishPackageVersionRequest"
      responses:
        "200":
          description: Version published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishPackageVersionResponse"
        "400":
          description: Invalid manifest or request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid user context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Package not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Version already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Database not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PackageSummary:
      type: object
      required: [id, name, visibility]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        visibility:
          type: string
          enum: [private, public]
        description:
          type: string
          nullable: true
    ListPackagesResponse:
      type: object
      required: [packages]
      properties:
        packages:
          type: array
          items:
            $ref: "#/components/schemas/PackageSummary"
    CreatePackageRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: hello-skill
        visibility:
          type: string
          enum: [private, public]
          default: private
        description:
          type: string
          nullable: true
    CreatePackageResponse:
      type: object
      required: [package]
      properties:
        package:
          $ref: "#/components/schemas/PackageSummary"
    PublishPackageVersionRequest:
      type: object
      required: [manifest]
      properties:
        manifest:
          type: object
          additionalProperties: true
    PublishPackageVersionResponse:
      type: object
      required: [package, version, artifact_digest]
      properties:
        package:
          type: string
        version:
          type: string
        artifact_digest:
          type: string
    PackageVersionSummary:
      type: object
      required: [version, artifact_digest, published_at]
      properties:
        version:
          type: string
        artifact_digest:
          type: string
        published_at:
          type: string
        deprecated_at:
          type: string
          nullable: true
    ListPackageVersionsResponse:
      type: object
      required: [package, versions]
      properties:
        package:
          type: string
        versions:
          type: array
          items:
            $ref: "#/components/schemas/PackageVersionSummary"
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string